Starting on #1306
