+++ db/repos/transactions.multisignature.js
@@ -14,26 +14,28 @@
 'use strict';
 
 var _ = require('lodash');
-var transactionTypes = require('../helpers/transactionTypes');
+var transactionTypes = require('../../helpers/transactionTypes');
 var columnSet;
 
 /**
- * InTransfer Transactions database interaction module
- * @memberof module:dapps
+ * Multisignature Transactions database interaction module
+ * @memberof module:multisignatures
  * @class
  * @param {Database} db - Instance of database object from pg-promise
  * @param {Object} pgp - pg-promise instance to utilize helpers
  * @constructor
- * @return {InTransferTransactionsRepo}
+ * @return {MultiSigTransactionsRepo}
  */
-function InTransferTransactionsRepo (db, pgp) {
+function MultiSigTransactionsRepo (db, pgp) {
 	this.db = db;
 	this.pgp = pgp;
 
-	this.dbTable = 'intransfer';
+	this.dbTable = 'multisignatures';
 
 	this.dbFields = [
-		'dappId',
+		'min',
+		'lifetime',
+		'keysgroup',
 		'transactionId'
 	];
 
@@ -47,18 +49,20 @@ function InTransferTransactionsRepo (db, pgp) {
 }
 
 /**
- * Save InTransfer transactions
- * @param {Array.<{id: string, asset: {inTransfer: {dappId: string}}}>} transactions
+ * Save Multisignature transactions
+ * @param {Array.<{id: string, {asset: {multisignature: {min: int, lifetime: int, keysgroup: Array.<string>}}}>} transactions
  * @return {Promise}
  */
-InTransferTransactionsRepo.prototype.save = function (transactions) {
+MultiSigTransactionsRepo.prototype.save = function (transactions) {
 	if (!_.isArray(transactions)) {
 		transactions = [transactions];
 	}
 
 	transactions = transactions.map(function (transaction) {
 		return {
-			dappId: transaction.asset.inTransfer.dappId,
+			min: transaction.asset.multisignature.min,
+			lifetime: transaction.asset.multisignature.lifetime,
+			keysgroup: transaction.asset.multisignature.keysgroup.join(','),
 			transactionId: transaction.id
 		};
 	});
@@ -66,4 +70,4 @@ InTransferTransactionsRepo.prototype.save = function (transactions) {
 	return this.db.none(this.pgp.helpers.insert(transactions, this.cs.insert));
 };
 
-module.exports = InTransferTransactionsRepo;
+module.exports = MultiSigTransactionsRepo;
