+++ test/common/DBSandbox.js
@@ -13,7 +13,7 @@
  */
 var child_process = require('child_process');
 
-var database = require('../../helpers/database.js');
+var database = require('../../db');
 
 var testDatabaseNames = [];
